UPSTREAM: soc/mediatek/mt8173/spi.c: Change assert to if statement
Asserts are only fatal if CONFIG_FATAL_ASSERTS is enabled in Kconfig.
By default this is disabled, so the assert is generally just a printf.

Die if someone decides to pass in an invalid bus number for some reason.

Addresses coverity issue 1349858 - Out-of-bounds read

Signed-off-by: Martin Roth <martinroth@google.com>
BUG=None
BRANCH=None
TEST=None

Reviewed-on: https://review.coreboot.org/17484
Reviewed-by: Nico Huber <nico.h@gmx.de>

Change-Id: I9d79bc336cbbfde31f655cfd271f101e7a90ab1b
Reviewed-on: https://chromium-review.googlesource.com/415068
Commit-Ready: Aaron Durbin <adurbin@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Aaron Durbin <adurbin@chromium.org>